abstact forms as client compoennts and pages as rscs
