+++ packages/generator/templates/app/src/app/auth/components/SignupForm.tsx
@@ -1,8 +1,10 @@
-import { LabeledTextField } from "src/app/components/LabeledTextField"
-import { Form, FORM_ERROR } from "src/app/components/Form"
+"use client"
+import {LabeledTextField} from "src/app/components/LabeledTextField"
+import {Form, FORM_ERROR} from "src/app/components/Form"
 import signup from "../mutations/signup"
-import { Signup } from "../validations"
-import { useMutation } from "@blitzjs/rpc"
+import {Signup} from "../validations"
+import {useMutation} from "@blitzjs/rpc"
+import {useRouter} from "next/navigation"
 
 type SignupFormProps = {
   onSuccess?: () => void
@@ -10,6 +12,8 @@ type SignupFormProps = {
 
 export const SignupForm = (props: SignupFormProps) => {
   const [signupMutation] = useMutation(signup)
+  const router = useRouter()
+
   return (
     <div>
       <h1>Create an Account</h1>
@@ -17,17 +21,18 @@ export const SignupForm = (props: SignupFormProps) => {
       <Form
         submitText="Create Account"
         schema={Signup}
-        initialValues={{ email: "", password: "" }}
+        initialValues={{email: "", password: ""}}
         onSubmit={async (values) => {
           try {
             await signupMutation(values)
-            props.onSuccess?.()
+            router.refresh()
+            router.push("/")
           } catch (error: any) {
             if (error.code === "P2002" && error.meta?.target?.includes("email")) {
               // This error comes from Prisma
-              return { email: "This email is already being used" }
+              return {email: "This email is already being used"}
             } else {
-              return { [FORM_ERROR]: error.toString() }
+              return {[FORM_ERROR]: error.toString()}
             }
           }
         }}
@@ -38,5 +43,3 @@ export const SignupForm = (props: SignupFormProps) => {
     </div>
   )
 }
-
-export default SignupForm
