+++ packages/generator/templates/app/src/app/auth/login/page.tsx
@@ -1,25 +1,5 @@
-"use client"
-import {useRouter} from "next/navigation"
 import {LoginForm} from "../components/LoginForm"
-import {useSearchParams} from "next/navigation"
-import { Route } from "next"
 
-const LoginPage = () => {
-  const router = useRouter()
-  const next = useSearchParams()?.get("next")
-  return (
-    <LoginForm
-      onSuccess={() => {
-        // refetch data for server components
-        router.refresh()
-        if (next) {
-          router.push(next as Route)
-        } else {
-          router.push("/")
-        }
-      }}
-    />
-  )
+export default function LoginPage() {
+  return <LoginForm />
 }
-
-export default LoginPage
